requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cryptography==2.5
Django==2.1.7
idna==2.8
jsonpickle==1.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
